refactor: DM redis 관련 파일 이동
- global -> domain/dm
